requirements.txt
Flask==2.0.1
face-recognition==1.3.0
Pillow==9.0.0
numpy==1.22.0
qrcode==7.3.1
Werkzeug==2.0.2
gunicorn==20.1.0
pymongo==4.3.3
dnspython==2.3.0
gevent>=1.4
